Separate JamsRadar into it's own file.
